LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-010-expected.txt


FAIL Should use the last fallback position initially assert_equals: Anchored element should be at the bottom of anchor expected -15 but got 485
FAIL Should use the third fallback position with enough space right assert_equals: Anchored element should be at the right of anchor expected 650 but got 0
FAIL Should use the second fallback position with enough space above assert_equals: Anchored element should be at the left of anchor expected 650 but got 100
FAIL Should use the first fallback position with enough space right and above assert_equals: Anchored element should be at the right of anchor expected 650 but got 0
